Refactors all customers ajax call and functions to add data to DOM
