+++ lib/objects/filter.js
@@ -1,7 +1,19 @@
 const $ = require('jquery')
 import { merchantRequests } from '../ajax-requests/merchant-requests'
-
+import { customerRequests } from '../ajax-requests/customer-requests'
 class Filter {
+
+  determineTableToLoad(event) {
+    let dataType = event.currentTarget.innerHTML
+    $('p.data-type').text(dataType)
+    if (dataType === 'Customers') {
+      filter.removeTableData([1])
+      customerRequests.getAllCustomers()
+    } else if (dataType === 'Merchants') {
+      filter.removeTableData([1])
+      merchantRequests.getAllMerchants()
+    }
+  }
 
   appendFilterResults() {
     let inputData = parseInt($('input').val())
